+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.03](https://github.com/arez/arez-spytools/tree/v0.03) (2018-04-27)
+[Full Changelog](https://github.com/arez/arez-spytools/compare/v0.02...v0.03)
 
 Update the 'org.realityforge.arez' dependencies to version '0.81'